+++ changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 2.0.0
-* 1.20.1!
+* Updated to 1.20.1
 * Added `/semihardcore <player> optout <true/false>` to opt out of the lives and max health changes 
 * When out of lives, dying now instantly respawns you at the same location where you died
 * Players are now informed when max health is lost